structural: extract ProducerLifecycleService from HlsStreamingService
Move the producer-mutation machinery (startTranscodes, resume, relocate,
suspend, mutex helpers) verbatim out of HlsStreamingService and
SessionReaper into a new ProducerLifecycleService that owns the
per-session mutex. HlsStreamingService and SessionReaper now delegate;
call order and locking are unchanged.

TranscodeRequest and TranscodeHandle gain an attemptId identifying one
producer run; it is plumbed through the engine, both executors, and the
fakes but read by nothing yet. Handle status transitions go through
withStatus, which preserves the attempt identity.

Prepares issue #252: recovery policy will drive replaceProducer on this
service while the coordinator owns the wait loop.
