+++ src/main/java/com/streamarr/server/services/streaming/SessionReaper.java
@@ -2,7 +2,6 @@ package com.streamarr.server.services.streaming;
 
 import com.streamarr.server.config.StreamingProperties;
 import com.streamarr.server.domain.streaming.StreamSession;
-import com.streamarr.server.domain.streaming.TranscodeHandle;
 import com.streamarr.server.domain.streaming.TranscodeStatus;
 import java.time.Instant;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +18,7 @@ public class SessionReaper {
   private final TranscodeExecutor transcodeExecutor;
   private final StreamingProperties properties;
   private final RuntimeStreamSessionRegistry runtimeRegistry;
+  private final ProducerLifecycleService producerLifecycle;
 
   @Scheduled(fixedDelayString = "${streaming.reaper-interval-ms:15000}")
   public void reapSessions() {
@@ -41,7 +41,7 @@ public class SessionReaper {
 
     if (isIdle(session, now) && session.hasActiveTranscodes()) {
       log.info("Suspending idle session {}", session.getSessionId());
-      suspendSession(session);
+      producerLifecycle.suspend(session);
       return;
     }
 
@@ -58,22 +58,6 @@ public class SessionReaper {
     return idleSeconds > properties.sessionTimeout().toSeconds();
   }
 
-  private void suspendSession(StreamSession session) {
-    transcodeExecutor.stop(session.getSessionId());
-    for (var entry : session.getVariantHandles().entrySet()) {
-      var handle = entry.getValue();
-      if (handle.status() != TranscodeStatus.ACTIVE) {
-        continue;
-      }
-
-      session.setVariantHandle(
-          entry.getKey(),
-          new TranscodeHandle(
-              handle.processId(), TranscodeStatus.SUSPENDED, handle.startSequenceNumber()));
-    }
-    runtimeRegistry.save(session);
-  }
-
   private void handleDeadProcesses(StreamSession session) {
     for (var entry : session.getVariantHandles().entrySet()) {
       var label = entry.getKey();
@@ -88,10 +72,7 @@ public class SessionReaper {
       }
 
       log.warn("FFmpeg process died for session {} variant {}", session.getSessionId(), label);
-      session.setVariantHandle(
-          label,
-          new TranscodeHandle(
-              handle.processId(), TranscodeStatus.FAILED, handle.startSequenceNumber()));
+      session.setVariantHandle(label, handle.withStatus(TranscodeStatus.FAILED));
       runtimeRegistry.save(session);
     }
   }
