+++ src/test/java/com/streamarr/server/fakes/FakeTranscodeExecutor.java
@@ -30,7 +30,8 @@ public class FakeTranscodeExecutor implements TranscodeExecutor {
     var key = new ProcessKey(request.sessionId(), request.variantLabel());
     running.add(key);
     started.add(key);
-    return new TranscodeHandle(1L, TranscodeStatus.ACTIVE, request.startSequenceNumber());
+    return new TranscodeHandle(
+        1L, request.attemptId(), TranscodeStatus.ACTIVE, request.startSequenceNumber());
   }
 
   @Override
