+++ src/test/java/com/streamarr/server/services/streaming/ProducerLifecycleServiceTest.java
@@ -1,0 +1,357 @@
+package com.streamarr.server.services.streaming;
+
+import static com.streamarr.server.fixtures.StreamSessionFixture.defaultSessionBuilder;
+import static com.streamarr.server.fixtures.StreamSessionFixture.defaultVariantBuilder;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
+
+import com.streamarr.server.config.StreamingProperties;
+import com.streamarr.server.domain.streaming.StreamSession;
+import com.streamarr.server.domain.streaming.TranscodeHandle;
+import com.streamarr.server.domain.streaming.TranscodeRequest;
+import com.streamarr.server.domain.streaming.TranscodeStatus;
+import com.streamarr.server.fakes.FakeRuntimeStreamSessionRegistry;
+import com.streamarr.server.fakes.FakeSegmentStore;
+import com.streamarr.server.fakes.FakeTranscodeExecutor;
+import com.streamarr.server.services.concurrency.MutexFactory;
+import java.time.Duration;
+import java.time.Instant;
+import java.util.List;
+import java.util.UUID;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+
+@Tag("UnitTest")
+@DisplayName("Producer Lifecycle Service Tests")
+class ProducerLifecycleServiceTest {
+
+  private FakeTranscodeExecutor transcodeExecutor;
+  private FakeSegmentStore segmentStore;
+  private FakeRuntimeStreamSessionRegistry runtimeRegistry;
+  private ProducerLifecycleService lifecycle;
+
+  @BeforeEach
+  void setUp() {
+    transcodeExecutor = new FakeTranscodeExecutor();
+    segmentStore = new FakeSegmentStore();
+    runtimeRegistry = new FakeRuntimeStreamSessionRegistry();
+    lifecycle =
+        ProducerLifecycleService.builder()
+            .transcodeExecutor(transcodeExecutor)
+            .segmentStore(segmentStore)
+            .properties(
+                StreamingProperties.builder()
+                    .maxConcurrentTranscodes(3)
+                    .targetSegmentDuration(Duration.ofSeconds(6))
+                    .sessionTimeout(Duration.ofSeconds(60))
+                    .build())
+            .runtimeRegistry(runtimeRegistry)
+            .sessionMutex(new MutexFactory<>())
+            .build();
+  }
+
+  private StreamSession startedSession() {
+    var session = defaultSessionBuilder().build();
+    runtimeRegistry.save(session);
+    lifecycle.startAll(session, 0, 0);
+    return session;
+  }
+
+  private StreamSession startedAbrSession() {
+    var session =
+        defaultSessionBuilder()
+            .variants(
+                List.of(
+                    defaultVariantBuilder()
+                        .width(1920)
+                        .height(1080)
+                        .videoBitrate(5_000_000L)
+                        .label("1080p")
+                        .build(),
+                    defaultVariantBuilder()
+                        .width(1280)
+                        .height(720)
+                        .videoBitrate(3_000_000L)
+                        .label("720p")
+                        .build()))
+            .build();
+    runtimeRegistry.save(session);
+    lifecycle.startAll(session, 0, 0);
+    return session;
+  }
+
+  private void suspendHandle(StreamSession session) {
+    session.setHandle(new TranscodeHandle(1L, TranscodeStatus.SUSPENDED));
+    transcodeExecutor.markDead(session.getSessionId());
+  }
+
+  @Test
+  @DisplayName("Should keep previously transcoded segments when relocating")
+  void shouldKeepPreviouslyTranscodedSegmentsWhenRelocating() {
+    var session = startedSession();
+    segmentStore.addSegment(session.getSessionId(), "segment0.ts", new byte[] {1});
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment100.ts");
+
+    // Segments are addressed on the absolute timeline, so earlier segments stay valid.
+    assertThat(segmentStore.segmentExists(session.getSessionId(), "segment0.ts")).isTrue();
+  }
+
+  @Test
+  @DisplayName("Should restart FFmpeg when segment is missing from suspended session")
+  void shouldRestartFfmpegWhenSegmentIsMissingFromSuspendedSession() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment5.ts");
+
+    assertThat(transcodeExecutor.isRunning(session.getSessionId())).isTrue();
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.ACTIVE);
+  }
+
+  @Test
+  @DisplayName("Should not restart FFmpeg when session is actively transcoding")
+  void shouldNotRestartFfmpegWhenSessionIsActivelyTranscoding() {
+    var session = startedSession();
+    var startedBefore = transcodeExecutor.getStartedRequests().size();
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment0.ts");
+
+    assertThat(transcodeExecutor.getStartedRequests()).hasSize(startedBefore);
+  }
+
+  @Test
+  @DisplayName("Should not restart FFmpeg when segment already exists on disk")
+  void shouldNotRestartFfmpegWhenSegmentAlreadyExistsOnDisk() {
+    var session = startedSession();
+    suspendHandle(session);
+    segmentStore.addSegment(session.getSessionId(), "segment5.ts", new byte[] {0x47});
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment5.ts");
+
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.SUSPENDED);
+  }
+
+  @Test
+  @DisplayName("Should not throw when positioning nonexistent session")
+  void shouldNotThrowWhenPositioningNonexistentSession() {
+    assertThatNoException()
+        .isThrownBy(() -> lifecycle.ensurePositioned(UUID.randomUUID(), "segment0.ts"));
+  }
+
+  @Test
+  @DisplayName("Should update last accessed time when resuming suspended session")
+  void shouldUpdateLastAccessedTimeWhenResumingSuspendedSession() {
+    var session = startedSession();
+    suspendHandle(session);
+    session.setLastAccessedAt(Instant.now().minusSeconds(200));
+    var oldAccessTime = session.getLastAccessedAt();
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment5.ts");
+
+    assertThat(session.getLastAccessedAt()).isAfter(oldAccessTime);
+  }
+
+  @Test
+  @DisplayName("Should resume with correct start number when segment is TS format")
+  void shouldResumeWithCorrectStartNumberWhenSegmentIsTsFormat() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment5.ts");
+
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.ACTIVE);
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.startSequenceNumber()).isEqualTo(5);
+    assertThat(lastRequest.seekPosition()).isEqualTo(30);
+  }
+
+  @Test
+  @DisplayName("Should resume with correct start number when segment is fMP4 format")
+  void shouldResumeWithCorrectStartNumberWhenSegmentIsFmp4Format() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment12.m4s");
+
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.ACTIVE);
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.startSequenceNumber()).isEqualTo(12);
+    assertThat(lastRequest.seekPosition()).isEqualTo(72);
+  }
+
+  @Test
+  @DisplayName("Should resume with correct start number when segment includes variant path")
+  void shouldResumeWithCorrectStartNumberWhenSegmentIncludesVariantPath() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "720p/segment3.ts");
+
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.ACTIVE);
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.startSequenceNumber()).isEqualTo(3);
+    assertThat(lastRequest.seekPosition()).isEqualTo(18);
+  }
+
+  @Test
+  @DisplayName("Should resume at beginning when segment name has no index")
+  void shouldResumeAtBeginningWhenSegmentNameHasNoIndex() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "init.mp4");
+
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.ACTIVE);
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.startSequenceNumber()).isZero();
+    assertThat(lastRequest.seekPosition()).isZero();
+  }
+
+  @Test
+  @DisplayName("Should resume at beginning when segment is first")
+  void shouldResumeAtBeginningWhenSegmentIsFirst() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment0.ts");
+
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.ACTIVE);
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.startSequenceNumber()).isZero();
+    assertThat(lastRequest.seekPosition()).isZero();
+  }
+
+  @Test
+  @DisplayName("Should restart all variant transcodes when ABR session is resumed")
+  void shouldRestartAllVariantTranscodesWhenAbrSessionIsResumed() {
+    var session = startedAbrSession();
+    var variantLabels = session.getVariants().stream().map(v -> v.label()).toList();
+
+    for (var label : variantLabels) {
+      session.setVariantHandle(label, new TranscodeHandle(1L, TranscodeStatus.SUSPENDED));
+      transcodeExecutor.markDead(session.getSessionId(), label);
+    }
+
+    var requestsBefore = transcodeExecutor.getStartedRequests().size();
+    lifecycle.ensurePositioned(session.getSessionId(), "segment5.ts");
+    var resumeRequests =
+        transcodeExecutor
+            .getStartedRequests()
+            .subList(requestsBefore, transcodeExecutor.getStartedRequests().size());
+
+    assertThat(resumeRequests).hasSize(variantLabels.size());
+    assertThat(resumeRequests).extracting(TranscodeRequest::startSequenceNumber).containsOnly(5);
+    assertThat(resumeRequests).extracting(TranscodeRequest::seekPosition).containsOnly(30);
+    assertThat(resumeRequests)
+        .extracting(TranscodeRequest::variantLabel)
+        .containsExactlyInAnyOrderElementsOf(variantLabels);
+    for (var label : variantLabels) {
+      assertThat(session.getVariantHandle(label).status()).isEqualTo(TranscodeStatus.ACTIVE);
+    }
+  }
+
+  @Test
+  @DisplayName(
+      "Should relocate the transcode when the requested segment is behind the encoder start")
+  void shouldRelocateTheTranscodeWhenTheRequestedSegmentIsBehindTheEncoderStart() {
+    var session = startedSession();
+    // Move the encoder forward first: segment50 is far ahead of fresh output.
+    lifecycle.ensurePositioned(session.getSessionId(), "segment50.ts");
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment10.ts");
+
+    // The encoder started at segment50 and will never produce segment10.
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.seekPosition()).isEqualTo(60);
+    assertThat(lastRequest.startSequenceNumber()).isEqualTo(10);
+  }
+
+  @Test
+  @DisplayName("Should relocate the transcode when the requested segment is far ahead of progress")
+  void shouldRelocateTheTranscodeWhenTheRequestedSegmentIsFarAheadOfProgress() {
+    var session = startedSession();
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment100.ts");
+
+    // Nothing near segment100 has been produced; waiting would stall the player.
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.seekPosition()).isEqualTo(600);
+    assertThat(lastRequest.startSequenceNumber()).isEqualTo(100);
+  }
+
+  @Test
+  @DisplayName("Should wait when the requested segment is near the encoder start")
+  void shouldWaitWhenTheRequestedSegmentIsNearTheEncoderStart() {
+    var session = startedSession();
+    var requestsBefore = transcodeExecutor.getStartedRequests().size();
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment2.ts");
+
+    // The encoder started at segment0 and will reach segment2 shortly.
+    assertThat(transcodeExecutor.getStartedRequests()).hasSize(requestsBefore);
+  }
+
+  @Test
+  @DisplayName("Should wait when the encoder is within the forward gap of the request")
+  void shouldWaitWhenTheEncoderIsWithinTheForwardGapOfTheRequest() {
+    var session = startedSession();
+    segmentStore.addSegment(session.getSessionId(), "segment96.ts", new byte[] {1});
+    var requestsBefore = transcodeExecutor.getStartedRequests().size();
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment100.ts");
+
+    // segment96 exists, so the encoder is close behind the request.
+    assertThat(transcodeExecutor.getStartedRequests()).hasSize(requestsBefore);
+  }
+
+  @Test
+  @DisplayName("Should not relocate when the requested segment already exists")
+  void shouldNotRelocateWhenTheRequestedSegmentAlreadyExists() {
+    var session = startedSession();
+    segmentStore.addSegment(session.getSessionId(), "segment10.ts", new byte[] {1});
+    var requestsBefore = transcodeExecutor.getStartedRequests().size();
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment10.ts");
+
+    assertThat(transcodeExecutor.getStartedRequests()).hasSize(requestsBefore);
+  }
+
+  @Test
+  @DisplayName("Should resume at the absolute segment position when resuming")
+  void shouldResumeAtTheAbsoluteSegmentPositionWhenResuming() {
+    var session = startedSession();
+    suspendHandle(session);
+
+    lifecycle.ensurePositioned(session.getSessionId(), "segment5.ts");
+
+    // The timeline is absolute: segment5 always covers [30s, 36s).
+    var lastRequest = transcodeExecutor.getStartedRequests().getLast();
+    assertThat(lastRequest.seekPosition()).isEqualTo(30);
+    assertThat(lastRequest.startSequenceNumber()).isEqualTo(5);
+  }
+
+  @Test
+  @DisplayName("Should stop producers and mark active handles suspended when suspending")
+  void shouldStopProducersAndMarkActiveHandlesSuspendedWhenSuspending() {
+    var session = startedSession();
+
+    lifecycle.suspend(session);
+
+    assertThat(transcodeExecutor.getStopped()).contains(session.getSessionId());
+    assertThat(session.getHandle().status()).isEqualTo(TranscodeStatus.SUSPENDED);
+  }
+
+  @Test
+  @DisplayName("Should preserve the attempt identity when suspending an active handle")
+  void shouldPreserveTheAttemptIdentityWhenSuspendingAnActiveHandle() {
+    var session = startedSession();
+    var attemptId = session.getHandle().attemptId();
+
+    lifecycle.suspend(session);
+
+    assertThat(session.getHandle().attemptId()).isEqualTo(attemptId);
+  }
+}
